PYIC-8992: migrate public credential jwk creator to javalin v7 instead of spark-core
Spark core hasn't been updated since 2022 and has had multiple vulnerabilities marked against it. Instead of using spark-core which isn't well-maintained, we're switching over to javalin which is better maintained and where the vulnerabilities are also resolved
