+++ public-jwk-creator/src/main/java/uk/gov/di/ipv/publiccredentialjwkcreator/App.java
@@ -1,0 +1,7 @@
+package uk.gov.di.ipv.publiccredentialjwkcreator;
+
+public class App {
+    public static void main(String[] args) {
+        new PublicCredentialJwkCreator();
+    }
+}
